Move `MessageSchema` to `stateful`
