+++ packages/engine/lib/stateful/src/message/mod.rs
@@ -1,0 +1,5 @@
+mod schema;
+
+pub use self::schema::{
+    MessageSchema, MESSAGE_ARROW_FIELDS, MESSAGE_BATCH_SCHEMA, MESSAGE_COLUMN_NAME,
+};
