+++ packages/engine/src/config/store.rs
@@ -1,10 +1,10 @@
 use std::sync::Arc;
 
-use stateful::agent::AgentSchema;
+use stateful::{agent::AgentSchema, message::MessageSchema};
 
 use crate::{
     config::{globals::Globals, Result},
-    datastore::schema::{context::ContextSchema, state::MessageSchema, EngineComponent},
+    datastore::schema::{context::ContextSchema, EngineComponent},
     simulation::package::creator::PackageCreators,
 };
