+++ packages/engine/src/datastore/schema/mod.rs
@@ -1,5 +1,4 @@
 pub mod context;
-pub mod state;
 
 mod field_spec;
