--- packages/engine/src/datastore/schema/state/mod.rs
@@ -1,3 +1,0 @@
-mod message;
-
-pub use self::message::MessageSchema;
